Debian support

+++ install-docker.sh
@@ -1,27 +1,35 @@
-#!/usr/bin/env bash 
+#!/usr/bin/env bash
 sudo apt update
 # (Install Docker CE)
 ## Set up the repository:
 ### Install packages to allow apt to use a repository over HTTPS
 sudo apt update && sudo apt install -y \
   apt-transport-https ca-certificates curl software-properties-common gnupg2
-  
+
 # Add Docker's official GPG key:
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo \
-  apt-key --keyring /etc/apt/trusted.gpg.d/docker.gpg add -
-  
-# Add the Docker apt repository:
-sudo add-apt-repository \
-  "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+if [ -f /etc/debian_version ]; then
+  curl -fsSL https://download.docker.com/linux/debian/gpg | sudo apt-key add -
+  sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/debian $(lsb_release -cs) stable"
+  sudo apt update
+  sudo apt install docker-ce
+  sudo usermod -aG docker ${USER}
+
+else
+  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo \
+    apt-key --keyring /etc/apt/trusted.gpg.d/docker.gpg add -
+
+  # Add the Docker apt repository:
+  sudo add-apt-repository \
+    "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
   $(lsb_release -cs) \
   stable"
-  
-# Install Docker CE
-sudo apt update && sudo apt install -y \
-  containerd.io=1.2.13-2 \
-  docker-ce=5:19.03.11~3-0~ubuntu-$(lsb_release -cs) \
-  docker-ce-cli=5:19.03.11~3-0~ubuntu-$(lsb_release -cs)
-  
+
+  # Install Docker CE
+  sudo apt update && sudo apt install -y \
+    containerd.io=1.2.13-2 \
+    docker-ce=5:19.03.11~3-0~ubuntu-$(lsb_release -cs) \
+    docker-ce-cli=5:19.03.11~3-0~ubuntu-$(lsb_release -cs)
+fi
 # Set up the Docker daemon
 cat <<EOF | sudo tee /etc/docker/daemon.json
 {
@@ -45,5 +53,3 @@ sudo systemctl enable docker
 
 echo "Docker installed"
 docker version
-
-
